+++ src/main/java/com/soma/backend/domain/chat/entity/ChatAttachment.java
@@ -9,5 +9,5 @@ public record ChatAttachment(
     String attachmentKey,
     String name,
     String contentType,
-    Long size) {
+    Integer size) {
 }
